rename and slightly restructure files to prepare for multiple context
